Fix bug (again) with `variable-usage` and `variable-declaration` (create an issue?) Move graph loading from .dot file to ImportUtils.kt Adjust corrupted graphs
